Rename config.py
