+++ scripts/conf.py
