+++ scripts/environment.py
@@ -1,6 +1,6 @@
 import os
 
-from scripts.config import BUILD_CLANG, BUILD_COMPILER_RT_MUSL, \
+from scripts.conf import BUILD_CLANG, BUILD_COMPILER_RT_MUSL, \
     BUILD_COMPILER_RT_NEWLIB, BUILD_LIBCXX, BUILD_LIBCXXABI, BUILD_LIBUNWIND, \
     BUILD_MUSL, BUILD_MUSL_HEADERS, BUILD_NEWLIB, BUILD_ROOT, MUSL_TRIPLE, \
     NEWLIB_TRIPLE, SRC_COMPILER_RT, SRC_LIBCXX, SRC_LIBCXXABI, SRC_LIBUNWIND, \
