+++ scripts/linux_headers.py
@@ -1,4 +1,4 @@
-from scripts.config import SRC_LINUX
+from scripts.conf import SRC_LINUX
 from scripts.record_step import is_linux_headers_configured, \
     is_linux_headers_installed, linux_headers_configured, \
     linux_headers_installed
